Archive/backend/app/schemas/auth_schema.py
# schema for Validation like joi/zod coz in-build
from pydantic import BaseModel, EmailStr


class UserCreate(BaseModel):
    email: EmailStr
    password: str


class UserLogin(BaseModel):
    email: EmailStr
    password: str