Archive/backend/app/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task queue
celery==5.4.0
redis==5.0.4

# Object storage
minio==7.2.7

# Search
elasticsearch==8.13.0

# OCR
paddlepaddle==2.6.1
paddleocr==2.7.3
opencv-python-headless==4.9.0.80
numpy==1.26.4

# File watching
watchdog==4.0.0

# Utils
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0